+++ src/main/java/participant/Players.java
@@ -1,8 +1,6 @@
 package participant;
 
 import card.Deck;
-import result.PlayerResult;
-import result.GameStatus;
 import card.Hand;
 import java.util.Collections;
 import java.util.List;
