+++ src/main/java/result/GameResult.java
@@ -3,6 +3,10 @@ package result;
 import java.util.List;
 import participant.Dealer;
 import participant.Player;
+import result.handStatus.BlackJack;
+import result.handStatus.Bust;
+import result.handStatus.HandStatus;
+import result.handStatus.ScoreComparison;
 
 public class GameResult {
     private final List<HandStatus> handStatuses;
