+++ src/main/java/result/handStatus/HandStatus.java
@@ -1,7 +1,8 @@
-package result;
+package result.handStatus;
 
 import participant.Dealer;
 import participant.Player;
+import result.GameStatus;
 
 public interface HandStatus {
     GameStatus calculateResult(Player player, Dealer dealer);
