+++ src/test/java/domain/participant/BettingTest.java
@@ -3,7 +3,7 @@ package domain.participant;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import participant.Money;
+import participant.value.Money;
 
 public class BettingTest {
     @DisplayName("베팅 금액을 정상적으로 지정한다")
